hubble,relay: use ca.crt from their own TLS Secret
Now that the CA certificate is stored in all the TLS Secrets as ca.crt,
we can use it for Hubble and Relay without needing to project the
cilium-ca Secret. This assume that the same CA is used for both
endpoints of the Hubble - Relay mTLS connection. Note that this was
already assumed before this patch, as both Hubble and Relay would use
the cilium-ca Secret as CA to authenticate the other party.

Signed-off-by: Alexandre Perrin <alex@kaworu.ch>