tests: Track Cilium pods as PodContexts, apply policies in all clusters
List Cilium nodes once in the beginning of tests and track them as
PodContext so that policy revision commands can be addressed to the
cluster the Cilium pod is running in.

Apply Cilium policy changes in all clusters by enumerating k8s clients
and clean up console output now that the same policies are applied in
multiple clusters.

Add ingress flow validation for TCP and use it in policy that only
allows one of the clients to ingress echo service. This verifies that
ingress policy enforcement works in multi-cluster scenario, as in that
case the echo-other-node pod is run in the destination cluster.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>