connectivity: package overhaul
Some of the goals of this refactor:

Overall separation of concerns, slicing the package into more clearly-defined
layers: ConnectivityTest -> []Test -> []Scenario -> []Action.

Move from an imperative (scripting) style to a declarative style.
A Test is now a scope, policies can be registered to it, and the
framework automatically takes care of applying and removing them
when appropriate. (including when interrupting the test program)
Tests that depend on state left behind from previous tests has
proven to be a source of bugs.

Providing a friendly test suite API. It should be abundantly clear which
policies are active during a test, and what the expected outcome is.

Improving output readability. Like Go's `testing`, if everything's peachy,
there doesn't need to be a flood of output. A lot of time was spent improving
this, but there's still lots to do.

- Test and Action now have their own Log(), Fail(), Fatal(), Debug() logging APIs:
  - Fail() marks the current test as failed.
  - Fatal() does Fail() and exits the current goroutine, firing defers.
- Within the scope of a Test, all output is buffered until a failure is
  encountered, after which it is flushed to stdout.
- The --verbose log level disables this buffering.
- A --debug log level was added for development. Debug statements were added
  in a few places, like command stdout.
- Logs generated at different levels are visually distinct.
  (e.g. Action logs are indented)

Simplifying the internal implementation and improving code readability.
We're expecting to add hundreds more tests, so maintainability is key.
Most interfaces that only had a single implementation were removed,
as I didn't find them to add any value (on the contrary - they got in the way).
A more canonical Go style was adopted instead: take interfaces, return structs.

Adding context cancellation to most of the infrastructure (addresses part of #67).
This sets us up for parallellizing tests. Each test now gets its own goroutine,
although only one executes at a time for now.

check.go was split out.

Signed-off-by: Timo Beckers <timo@isovalent.com>