getFlows: support deadline flow returns
previous to this commit an error would be returned
with no flows if a Deadline context was provided
to the target method.

so follow the same logic as a Cancel context being passed
in, the switch statement is updated to also return flows
if a Deadline is exceeded.

in this new case all flows received before the Deadline
will be returned.

Signed-off-by: ldelossa <louis.delos@isovalent.com>