k8s: Check observedGeneration in status
When a DaemonSet or Deployment is updated, it might take a while for the
status field to reflect the status of the updated version. If we check
the status too soon, there is a high chance that we are checking an
outdated status.

Therefore, check the observedGeneration field first, before reading out
the status flag. This ensures that the reported status actually refers
to the most recent desired state.

Recommended reading: https://alenkacz.medium.com/-250728868792

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>