Switch pod to world testing from `cilium.io` to `one.one.one.one`
Tentative fix for #367.

We have two main hypothesis around why connections to external domains
are flaky:

- The external domain itself is unreliable.
- The external domain works fine, but CoreDNS is unreliable.

`cilium.io` is hosted on single non-HA EC2 instance, which is definitely
the most robus thing out there. We propose switching to
`one.one.one.one` as a quick fix to check if that helps with the
reliability.

If yes, we will evaluate moving to a HA system that we control ourselves
(e.g. a DNS zone hosted at a major provider).

If not, we will investigate the CoreDNS hypothesis.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>