add new `aks-byocni` datapath mode
Add a new datapath mode for AKS clusters created using the new BYOCNI
capabilities. This datapath mode is mutually exclusive with the older
Azure datapath mode, as Azure IPAM is not available when using BYOCNI.

A similar datapath mode is going to be added to the main Helm chart so
that enabling Cilium for AKS BYOCNI clusters is as simple as setting
`aksbyocni.enabled=true` in the Helm values, in a similar fashion to
what is done for the older Azure integration via `azure.enabled=true.

However, since this new integration has not yet landed in stable
releases, we temporarily manually configure the same ConfigMap options
as the new `aksbyocni` mode does, so that it works transparently.

See main PR on `cilium/cilium` for more context and details:
https://github.com/cilium/cilium/pull/19379

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>